Bot collection now contains bot from RLBotGUI's RLBotPack if present.
